department

+++ 01SOURCE/base-common-web/src/main/java/cn/doublepoint/common/port/adapter/template/controller/menu/MenuManagementController.java
@@ -66,7 +66,7 @@ public class MenuManagementController extends BaseController {
 		String id = request.getParameter("id");
 		String menuName = "";
 		try {
-			SysMenu menu = menuService.getById(Long.valueOf(id));
+			SysMenu menu = menuService.getById(Integer.valueOf(id));
 			if (menu != null) {
 				menuName = menu.getName();
 			}
