+++ 01SOURCE/base-common-web/src/main/java/cn/doublepoint/common/port/adapter/template/controller/menu/MenuTreeController.java
@@ -39,7 +39,7 @@ public class MenuTreeController extends BaseTreeController {
 	@ResponseBody
 	public String getMenuTree(@RequestParam(required=true)String code) {
 		try{
-			SysMenu menu = menuService.getById(Long.valueOf(code));
+			SysMenu menu = menuService.getById(Integer.valueOf(code));
 			return menu.getName();
 		}
 		catch(Exception e){
